Make running problems more ergonomic

+++ src/problems/p07.rs
@@ -1,7 +1,7 @@
 //use crate::primes;
 
-pub fn p07() -> u32 {
-    crate::primes::get_primes()[10_000]
+pub fn p07() -> u64 {
+    crate::primes::get_primes()[10_000] as u64
 }
 
 #[cfg(test)]
